Add ModuleTypeWidgetLibraryWrapper, Move ModuleContext to engine.module
